% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_pull_request.R
\name{examine_pull_request}
\alias{examine_pull_request}
\title{Examine pull request}
\usage{
examine_pull_request(user, repo, branch, default_branch = "main")
}
\arguments{
\item{user}{GitHub user name}

\item{repo}{GitHub repo name (the name of the fork if this differs from the name of the original repo)}

\item{branch}{name of branch to examine - this will be given in pull request text}

\item{default_branch}{name of default branch (probably main or master)}
}
\description{
Examine pull request
}
\examples{
#examine_pull_request(user = "audhalbritter", repo = "biostats", branch = "FixGitTutorial") 
}
